Doing "ReactApp - Tic Tac Toe": - I have created a new service for user needs and separated   responsibilities from Login and Users - Created new actions and reducers for User services and Store
