+++ react/react-training/src/redux/login/reducer.js
@@ -11,21 +11,6 @@ const initialState = {
 
 const Reducer = (state = initialState, action) => {
   switch (action.type) {
-    case actionTypes.USER_GET:
-      return {
-        ...state,
-        ...action.payload
-      };
-    case actionTypes.USER_GET_FAILURE:
-      return {
-        ...state,
-        ...action.payload
-      };
-    case actionTypes.USER_GET_SUCCESS:
-      return {
-        ...state,
-        ...action.payload
-      };
     case actionTypes.USER_LOGIN:
       return {
         ...state,
@@ -37,36 +22,6 @@ const Reducer = (state = initialState, action) => {
         ...action.payload
       };
     case actionTypes.USER_LOGIN_SUCCESS:
-      return {
-        ...state,
-        ...action.payload
-      };
-    case actionTypes.USER_PATCH:
-      return {
-        ...state,
-        ...action.payload
-      };
-    case actionTypes.USER_PATCH_FAILURE:
-      return {
-        ...state,
-        ...action.payload
-      };
-    case actionTypes.USER_PATCH_SUCCESS:
-      return {
-        ...state,
-        ...action.payload
-      };
-    case actionTypes.USER_POST:
-      return {
-        ...state,
-        ...action.payload
-      };
-    case actionTypes.USER_POST_FAILURE:
-      return {
-        ...state,
-        ...action.payload
-      };
-    case actionTypes.USER_POST_SUCCESS:
       return {
         ...state,
         ...action.payload
