+++ react/react-training/src/redux/store.js
@@ -1,5 +1,6 @@
 import GameReducer from '@redux/game/reducer';
 import LoginReducer from '@redux/login/reducer';
+import UserReducer from '@redux/user/reducer';
 import thunk from 'redux-thunk';
 import { applyMiddleware, combineReducers, compose, createStore } from 'redux';
 import { reducer as formReducer } from 'redux-form';
@@ -7,6 +8,7 @@ import { reducer as formReducer } from 'redux-form';
 const reducers = combineReducers({
   GameReducer,
   LoginReducer,
+  UserReducer,
   form: formReducer
 });
