+++ react/react-training/src/services/LoginService.js
@@ -1,0 +1,5 @@
+import api from '../config/api';
+
+export default {
+  userLogin: (username, password) => api.post('/Users/login', { username, password })
+};
