+++ react/react-training/src/services/UserService.js
@@ -2,7 +2,6 @@ import api from '../config/api';
 
 export default {
   userGet: userId => api.get(`/Users/${userId}`),
-  userLogin: (username, password) => api.post('/Users/login', { username, password }),
   userPatch: (userId, newUserData) => api.patch(`/Users/${userId}`, newUserData),
   userPost: user => api.post('/Users', user)
 };
